resume-generator/requirements.txt
# 核心依赖
pandas>=1.3.0
openpyxl>=3.0.0  # Excel读写
python-docx>=0.8.11  # Word文档处理
reportlab>=3.6.0  # PDF生成
Jinja2>=3.0.0  # 模板引擎

# 工具依赖
PyYAML>=6.0  # 配置文件处理
tqdm>=4.60.0  # 进度条显示
click>=8.0.0  # 命令行界面

# 开发依赖
pytest>=6.0.0
pytest-cov>=2.0.0
black>=21.0.0
flake8>=3.9.0